requirements.txt
Django==1.7.7
django-appconf==1.0.1
django-crontab==0.6.0
django-nested-inline==0.3.4
django-suit==0.2.13
django-tastypie==0.12.1
django-tastypie-swagger==0.1.4
Pillow==2.8.2
psycopg2==2.6.1
python-dateutil==2.4.2
python-gcm==0.2
python-mimeparse==0.1.4
reportlab==3.2.0
requests==2.7.0
simplejson==3.7.3
six==1.9.0
urllib3==1.10.4
wheel==0.24.0
django-geoposition==0.2.2